requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.0
pandas
openpyxl
fpdf2
Werkzeug>=3.0.0
python-dotenv
gunicorn==21.2.0
supabase>=2.3.0
